refactor all commands cli to make it consistent.
