+++ pkg/cli/cli.go
@@ -4,41 +4,11 @@
 package cli
 
 import (
-	"context"
-	"math/big"
-	"os"
-	"syscall"
-	"time"
-
-	"github.com/ethereum/go-ethereum/accounts/abi/bind"
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/params"
-	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
-	"github.com/oklog/run"
 	"github.com/pkg/errors"
-	"github.com/prometheus/prometheus/storage"
-	"github.com/prometheus/prometheus/tsdb"
-	"github.com/tellor-io/telliot/pkg/aggregator"
 	"github.com/tellor-io/telliot/pkg/config"
-	"github.com/tellor-io/telliot/pkg/contracts"
-	"github.com/tellor-io/telliot/pkg/db"
 	"github.com/tellor-io/telliot/pkg/ethereum"
-	tEthereum "github.com/tellor-io/telliot/pkg/ethereum"
-	"github.com/tellor-io/telliot/pkg/gasPrice/gasStation"
 	"github.com/tellor-io/telliot/pkg/logging"
-	"github.com/tellor-io/telliot/pkg/mining"
-	psrTellor "github.com/tellor-io/telliot/pkg/psr/tellor"
-	psrTellorMesosphere "github.com/tellor-io/telliot/pkg/psr/tellorMesosphere"
-	"github.com/tellor-io/telliot/pkg/reward"
-	"github.com/tellor-io/telliot/pkg/submitter/tellor"
-	"github.com/tellor-io/telliot/pkg/submitter/tellorMesosphere"
-	"github.com/tellor-io/telliot/pkg/tasker"
-	"github.com/tellor-io/telliot/pkg/tracker/dispute"
-	"github.com/tellor-io/telliot/pkg/tracker/index"
-	"github.com/tellor-io/telliot/pkg/tracker/profit"
-	"github.com/tellor-io/telliot/pkg/transactor"
-	"github.com/tellor-io/telliot/pkg/web"
 )
 
 const VersionMessage = `
@@ -74,100 +44,38 @@ type VersionCmd struct {
 }
 
 func (cmd *VersionCmd) Run() error {
-	// The main entry point prints the version message so here just return nil and the message will be printed.
+	// The main entry point prints the version message so
+	// here just return nil and the message will be printed.
 	return nil
 }
 
+type cfg struct {
+	Config configPath `type:"existingfile" help:"path to config file"`
+}
+
+type addr struct {
+	Addr string `arg:"" required:""`
+}
+
+type cfgGas struct {
+	cfg
+	GasPrice int `optional:"" help:"gas price to use when running the command"`
+}
+
+type cfgGasAddr struct {
+	cfgGas
+	addr
+}
+
+type cfgAddr struct {
+	cfg
+	addr
+}
+
 type configPath string
-type tokenCmd struct {
-	Config   configPath `type:"existingfile" help:"path to config file"`
-	Address  string     `arg:""`
-	Amount   string     `arg:""`
-	Account  int        `optional:""`
-	GasPrice int        `optional:"" help:"gas price in gwei at which to run the command"`
-}
-
-type transferCmd tokenCmd
-
-func (self *transferCmd) Run() error {
-	logger := logging.NewLogger()
-	ctx := context.Background()
-
-	_, err := config.ParseConfig(logger, string(self.Config)) // Load the env file.
-	if err != nil {
-		return errors.Wrap(err, "creating config")
-	}
-
-	client, err := ethereum.NewClient(ctx, logger)
-	if err != nil {
-		return errors.Wrap(err, "creating ethereum client")
-	}
-
-	address := ETHAddress{}
-	err = address.Set(self.Address)
-	if err != nil {
-		return errors.Wrap(err, "parsing address argument")
-	}
-	amount := TRBAmount{}
-	err = amount.Set(self.Amount)
-	if err != nil {
-		return errors.Wrap(err, "parsing amount argument")
-	}
-	account, err := ethereum.GetAccountFor(self.Account)
-	if err != nil {
-		return err
-	}
-
-	contract, err := contracts.NewITellor(client)
-	if err != nil {
-		return errors.Wrap(err, "create tellor contract instance")
-	}
-
-	return Transfer(ctx, logger, client, contract, account, address.addr, amount.Int, self.GasPrice)
-
-}
-
-type approveCmd tokenCmd
-
-func (self *approveCmd) Run() error {
-	logger := logging.NewLogger()
-	ctx := context.Background()
-
-	_, err := config.ParseConfig(logger, string(self.Config)) // Load the env file.
-	if err != nil {
-		return errors.Wrap(err, "creating config")
-	}
-
-	client, err := ethereum.NewClient(ctx, logger)
-	if err != nil {
-		return errors.Wrap(err, "creating ethereum client")
-	}
-
-	address := ETHAddress{}
-	err = address.Set(self.Address)
-	if err != nil {
-		return errors.Wrap(err, "parsing address argument")
-	}
-	amount := TRBAmount{}
-	err = amount.Set(self.Amount)
-	if err != nil {
-		return errors.Wrap(err, "parsing amount argument")
-	}
-	account, err := ethereum.GetAccountFor(self.Account)
-	if err != nil {
-		return err
-	}
-
-	contract, err := contracts.NewITellor(client)
-	if err != nil {
-		return errors.Wrap(err, "create tellor contract instance")
-	}
-
-	return Approve(ctx, logger, client, contract, account, address.addr, amount.Int, self.GasPrice)
-}
 
 type accountsCmd struct {
-	Config configPath `type:"existingfile" help:"path to config file"`
+	cfg
 }
 
 func (self *accountsCmd) Run() error {
@@ -187,797 +95,5 @@ func (self *accountsCmd) Run() error {
 		level.Info(logger).Log("msg", "account", "no", i, "address", account.Address.String())
 	}
 
-	return nil
-}
-
-type balanceCmd struct {
-	Config  configPath `type:"existingfile" help:"path to config file"`
-	Address string     `arg:"" optional:""`
-}
-
-func (self *balanceCmd) Run() error {
-	logger := logging.NewLogger()
-	ctx := context.Background()
-
-	_, err := config.ParseConfig(logger, string(self.Config)) // Load the env file.
-	if err != nil {
-		return errors.Wrap(err, "creating config")
-	}
-
-	client, err := ethereum.NewClient(ctx, logger)
-	if err != nil {
-		return errors.Wrap(err, "creating ethereum client")
-	}
-
-	contract, err := contracts.NewITellor(client)
-	if err != nil {
-		return errors.Wrap(err, "create tellor contract instance")
-	}
-
-	addr := ETHAddress{}
-	if self.Address == "" {
-		err = addr.Set(contract.Address.String())
-		if err != nil {
-			return errors.Wrap(err, "parsing argument")
-		}
-	} else {
-		err = addr.Set(self.Address)
-		if err != nil {
-			return errors.Wrap(err, "parsing argument")
-		}
-	}
-
-	return Balance(ctx, logger, client, contract, addr.addr)
-}
-
-type depositCmd struct {
-	Config   configPath `type:"existingfile" help:"path to config file"`
-	Account  int        `optional:""`
-	GasPrice int        `optional:"" help:"gas price in gwei at which to run the command"`
-}
-
-func (self depositCmd) Run() error {
-	logger := logging.NewLogger()
-	ctx := context.Background()
-
-	_, err := config.ParseConfig(logger, string(self.Config)) // Load the env file.
-	if err != nil {
-		return errors.Wrap(err, "creating config")
-	}
-	client, err := ethereum.NewClient(ctx, logger)
-	if err != nil {
-		return errors.Wrap(err, "creating ethereum client")
-	}
-	account, err := ethereum.GetAccountFor(self.Account)
-	if err != nil {
-		return err
-	}
-	contract, err := contracts.NewITellor(client)
-	if err != nil {
-		return errors.Wrap(err, "create tellor contract instance")
-	}
-	return Deposit(ctx, logger, client, contract, account, self.GasPrice)
-
-}
-
-type withdrawCmd struct {
-	Config   configPath `type:"existingfile" help:"path to config file"`
-	Address  string     `arg:"" required:""`
-	Account  int        `optional:""`
-	GasPrice int        `optional:"" help:"gas price in gwei at which to run the command"`
-}
-
-func (self withdrawCmd) Run() error {
-	logger := logging.NewLogger()
-	ctx := context.Background()
-
-	_, err := config.ParseConfig(logger, string(self.Config)) // Load the env file.
-	if err != nil {
-		return errors.Wrap(err, "creating config")
-	}
-
-	client, err := ethereum.NewClient(ctx, logger)
-	if err != nil {
-		return errors.Wrap(err, "creating ethereum client")
-	}
-
-	addr := ETHAddress{}
-	err = addr.Set(self.Address)
-	if err != nil {
-		return errors.Wrap(err, "parsing argument")
-	}
-	account, err := ethereum.GetAccountFor(self.Account)
-	if err != nil {
-		return err
-	}
-	contract, err := contracts.NewITellor(client)
-	if err != nil {
-		return errors.Wrap(err, "create tellor contract instance")
-	}
-	return WithdrawStake(ctx, logger, client, contract, account, self.GasPrice)
-
-}
-
-type requestCmd struct {
-	Config   configPath `type:"existingfile" help:"path to config file"`
-	Account  int        `optional:""`
-	GasPrice int        `optional:"" help:"gas price in gwei at which to run the command"`
-}
-
-func (self requestCmd) Run() error {
-	logger := logging.NewLogger()
-	ctx := context.Background()
-
-	_, err := config.ParseConfig(logger, string(self.Config)) // Load the env file.
-	if err != nil {
-		return errors.Wrap(err, "creating config")
-	}
-
-	client, err := ethereum.NewClient(ctx, logger)
-	if err != nil {
-		return errors.Wrap(err, "creating ethereum client")
-	}
-	account, err := ethereum.GetAccountFor(self.Account)
-	if err != nil {
-		return err
-	}
-	contract, err := contracts.NewITellor(client)
-	if err != nil {
-		return errors.Wrap(err, "create tellor contract instance")
-	}
-	return RequestStakingWithdraw(ctx, logger, client, contract, account, self.GasPrice)
-}
-
-type statusCmd struct {
-	Config  configPath `type:"existingfile" help:"path to config file"`
-	Account int        `optional:""`
-}
-
-func (self statusCmd) Run() error {
-	logger := logging.NewLogger()
-	ctx := context.Background()
-
-	_, err := config.ParseConfig(logger, string(self.Config)) // Load the env file.
-	if err != nil {
-		return errors.Wrap(err, "creating config")
-	}
-
-	client, err := ethereum.NewClient(ctx, logger)
-	if err != nil {
-		return errors.Wrap(err, "creating ethereum client")
-	}
-	account, err := ethereum.GetAccountFor(self.Account)
-	if err != nil {
-		return err
-	}
-	contract, err := contracts.NewITellor(client)
-	if err != nil {
-		return errors.Wrap(err, "create tellor contract instance")
-	}
-	return ShowStatus(ctx, logger, client, contract, account)
-}
-
-type newDisputeCmd struct {
-	Config     configPath `type:"existingfile" help:"path to config file"`
-	RequestID  string     `arg:""  help:"the request id to dispute it"`
-	Timestamp  string     `arg:""  help:"the submitted timestamp to dispute"`
-	MinerIndex string     `arg:""  help:"the miner index to dispute"`
-	Account    int        `optional:""`
-	GasPrice   int        `optional:"" help:"gas price in gwei at which to run the command"`
-}
-
-func (self newDisputeCmd) Run() error {
-	logger := logging.NewLogger()
-	ctx := context.Background()
-
-	_, err := config.ParseConfig(logger, string(self.Config)) // Load the env file.
-	if err != nil {
-		return errors.Wrap(err, "creating config")
-	}
-
-	client, err := ethereum.NewClient(ctx, logger)
-	if err != nil {
-		return errors.Wrap(err, "creating ethereum client")
-	}
-
-	requestID := EthereumInt{}
-	err = requestID.Set(self.RequestID)
-	if err != nil {
-		return errors.Wrap(err, "parsing argument")
-	}
-	timestamp := EthereumInt{}
-	err = timestamp.Set(self.Timestamp)
-	if err != nil {
-		return errors.Wrap(err, "parsing argument")
-	}
-	minerIndex := EthereumInt{}
-	err = minerIndex.Set(self.MinerIndex)
-	if err != nil {
-		return errors.Wrap(err, "parsing argument")
-	}
-	account, err := ethereum.GetAccountFor(self.Account)
-	if err != nil {
-		return err
-	}
-	contract, err := contracts.NewITellor(client)
-	if err != nil {
-		return errors.Wrap(err, "create tellor contract instance")
-	}
-	return Dispute(ctx, logger, client, contract, account, requestID.Int, timestamp.Int, minerIndex.Int, self.GasPrice)
-}
-
-type voteCmd struct {
-	Config    configPath `type:"existingfile" help:"path to config file"`
-	disputeId string     `arg:"" required:"" help:"the dispute id"`
-	support   bool       `arg:""  help:"true or false"`
-	Account   int        `optional:""`
-	GasPrice  int        `optional:"" help:"gas price in gwei at which to run the command"`
-}
-
-func (self voteCmd) Run() error {
-	logger := logging.NewLogger()
-	ctx := context.Background()
-
-	_, err := config.ParseConfig(logger, string(self.Config)) // Load the env file.
-	if err != nil {
-		return errors.Wrap(err, "creating config")
-	}
-
-	client, err := ethereum.NewClient(ctx, logger)
-	if err != nil {
-		return errors.Wrap(err, "creating ethereum client")
-	}
-
-	disputeID := EthereumInt{}
-	err = disputeID.Set(self.disputeId)
-	if err != nil {
-		return errors.Wrap(err, "parsing argument")
-	}
-	account, err := ethereum.GetAccountFor(self.Account)
-	if err != nil {
-		return err
-	}
-	contract, err := contracts.NewITellor(client)
-	if err != nil {
-		return errors.Wrap(err, "create tellor contract instance")
-	}
-
-	voted, err := contract.DidVote(&bind.CallOpts{Context: ctx}, disputeID.Int, contract.Address)
-	if err != nil {
-		return errors.Wrapf(err, "check if you've already voted")
-	}
-	if voted {
-		level.Info(logger).Log("msg", "you have already voted on this dispute")
-		return nil
-	}
-
-	var gasPrice *big.Int
-	if self.GasPrice > 0 {
-		gasPrice = big.NewInt(int64(self.GasPrice) * params.GWei)
-	}
-
-	auth, err := tEthereum.PrepareEthTransaction(ctx, client, account, gasPrice)
-	if err != nil {
-		return errors.Wrapf(err, "prepare ethereum transaction")
-	}
-	tx, err := contract.Vote(auth, disputeID.Int, self.support)
-	if err != nil {
-		return errors.Wrapf(err, "submit vote transaction")
-	}
-
-	level.Info(logger).Log("msg", "vote submitted with transaction", "tx", tx.Hash().Hex())
-	return nil
-}
-
-type tallyCmd struct {
-	Config    configPath `type:"existingfile" help:"path to config file"`
-	DisputeID string     `arg:"" help:"the dispute id"`
-	Account   int        `optional:""`
-	GasPrice  int        `optional:"" help:"gas price in gwei at which to run the command"`
-}
-
-func (self tallyCmd) Run() error {
-	logger := logging.NewLogger()
-	ctx := context.Background()
-
-	_, err := config.ParseConfig(logger, string(self.Config)) // Load the env file.
-	if err != nil {
-		return errors.Wrap(err, "creating config")
-	}
-
-	client, err := ethereum.NewClient(ctx, logger)
-	if err != nil {
-		return errors.Wrap(err, "creating ethereum client")
-	}
-
-	disputeID := EthereumInt{}
-	err = disputeID.Set(self.DisputeID)
-	if err != nil {
-		return errors.Wrap(err, "parsing argument")
-	}
-	account, err := ethereum.GetAccountFor(self.Account)
-	if err != nil {
-		return err
-	}
-	contract, err := contracts.NewITellor(client)
-	if err != nil {
-		return errors.Wrap(err, "create tellor contract instance")
-	}
-
-	var gasPrice *big.Int
-	if self.GasPrice > 0 {
-		gasPrice = big.NewInt(int64(self.GasPrice) * params.GWei)
-	}
-
-	auth, err := tEthereum.PrepareEthTransaction(ctx, client, account, gasPrice)
-	if err != nil {
-		return errors.Wrapf(err, "prepare ethereum transaction")
-	}
-
-	tx, err := contract.TallyVotes(auth, disputeID.Int)
-	if err != nil {
-		return errors.Wrapf(err, "run tally votes if you've already voted")
-	}
-
-	level.Info(logger).Log("msg", "tally votes submitted", "tx", tx.Hash().Hex())
-	return nil
-}
-
-type listCmd struct {
-	Config  configPath `type:"existingfile" help:"path to config file"`
-	Account int        `optional:""`
-}
-
-func (self listCmd) Run() error {
-	logger := logging.NewLogger()
-	ctx := context.Background()
-
-	cfg, err := config.ParseConfig(logger, string(self.Config)) // Load the env file.
-	if err != nil {
-		return errors.Wrap(err, "creating config")
-	}
-
-	client, err := ethereum.NewClient(ctx, logger)
-	if err != nil {
-		return errors.Wrap(err, "creating ethereum client")
-	}
-	account, err := ethereum.GetAccountFor(self.Account)
-	if err != nil {
-		return err
-	}
-
-	// Open the TSDB database.
-	var querable storage.SampleAndChunkQueryable
-	if cfg.Db.RemoteHost != "" {
-		querable, err = db.NewRemoteDB(cfg.Db)
-		if err != nil {
-			return errors.Wrap(err, "opening remote tsdb DB")
-		}
-	} else {
-		if err := os.MkdirAll(cfg.Db.Path, 0777); err != nil {
-			return errors.Wrap(err, "creating tsdb DB folder")
-		}
-		tsdbOptions := tsdb.DefaultOptions()
-		tsdbOptions.NoLockfile = true
-		querable, err = tsdb.Open(cfg.Db.Path, nil, nil, tsdbOptions)
-		if err != nil {
-			return errors.Wrap(err, "opening tsdb DB")
-		}
-	}
-
-	aggregator, err := aggregator.New(logger, ctx, cfg.Aggregator, querable)
-	if err != nil {
-		return errors.Wrap(err, "creating aggregator")
-	}
-
-	psr := psrTellor.New(logger, cfg.PsrTellor, aggregator)
-	contract, err := contracts.NewITellor(client)
-	if err != nil {
-		return errors.Wrap(err, "create tellor contract instance")
-	}
-	return List(ctx, logger, client, contract, account, psr)
-}
-
-type dataserverCmd struct {
-	Config configPath `type:"existingfile" help:"path to config file"`
-}
-
-func (self dataserverCmd) Run() error {
-	logger := logging.NewLogger()
-
-	cfg, err := config.ParseConfig(logger, string(self.Config))
-	if err != nil {
-		return errors.Wrap(err, "creating config")
-	}
-
-	// Defining a global context for starting and stopping of components.
-	ctx := context.Background()
-
-	// We define our run groups here.
-	var g run.Group
-	// Run groups.
-	{
-		// Handle interupts.
-		g.Add(run.SignalHandler(context.Background(), syscall.SIGHUP, syscall.SIGINT, syscall.SIGTERM))
-
-		// Open the TSDB database.
-		tsdbOptions := tsdb.DefaultOptions()
-		// 5 days are enough as the aggregator needs data only 24 hours in the past.
-		tsdbOptions.RetentionDuration = int64(5 * 24 * time.Hour / time.Millisecond)
-		if err := os.MkdirAll(cfg.Db.Path, 0777); err != nil {
-			return errors.Wrap(err, "creating tsdb DB folder")
-		}
-		tsDB, err := tsdb.Open(cfg.Db.Path, nil, nil, tsdbOptions)
-		if err != nil {
-			return errors.Wrap(err, "creating tsdb DB")
-		}
-		level.Info(logger).Log("msg", "opened local db", "path", cfg.Db.Path)
-
-		defer func() {
-			if err := tsDB.Close(); err != nil {
-				level.Error(logger).Log("msg", "closing the tsdb", "err", err)
-			}
-		}()
-
-		// Index tracker.
-
-		// The client is needed when the api requests data from the blockchain.
-		// TODO create an eth client only if the api config file has eth address.
-		client, err := ethereum.NewClient(ctx, logger)
-		if err != nil {
-			return errors.Wrap(err, "creating ethereum client")
-		}
-
-		index, err := index.New(logger, ctx, cfg.IndexTracker, tsDB, client)
-		if err != nil {
-			return errors.Wrap(err, "creating index tracker")
-		}
-
-		g.Add(func() error {
-			err := index.Run()
-			level.Info(logger).Log("msg", "index shutdown complete")
-			return err
-		}, func(error) {
-			index.Stop()
-		})
-
-		// Aggregator.
-		aggregator, err := aggregator.New(logger, ctx, cfg.Aggregator, tsDB)
-		if err != nil {
-			return errors.Wrap(err, "creating aggregator")
-		}
-
-		contractTellor, err := contracts.NewITellor(client)
-		if err != nil {
-			return errors.Wrap(err, "create tellor contract instance")
-		}
-
-		disputeTracker, err := dispute.New(
-			logger,
-			ctx,
-			cfg.DisputeTracker,
-			tsDB,
-			client,
-			contractTellor,
-			psrTellor.New(logger, cfg.PsrTellor, aggregator),
-		)
-		if err != nil {
-			return errors.Wrap(err, "creating profit tracker")
-		}
-		g.Add(func() error {
-			disputeTracker.Start()
-			level.Info(logger).Log("msg", "dispute tracker shutdown complete")
-			return nil
-		}, func(error) {
-			disputeTracker.Stop()
-		})
-
-		// Web/Api server.
-		{
-			srv, err := web.New(logger, ctx, tsDB, cfg.Web)
-			if err != nil {
-				return errors.Wrap(err, "create web server")
-			}
-			g.Add(func() error {
-				err := srv.Start()
-				level.Info(logger).Log("msg", "web server shutdown complete")
-				return err
-			}, func(error) {
-				srv.Stop()
-			})
-		}
-	}
-
-	if err := g.Run(); err != nil {
-		level.Error(logger).Log("msg", "main exited with error", "err", err)
-		return err
-	}
-
-	level.Info(logger).Log("msg", "main shutdown complete")
-	return nil
-}
-
-type mineCmd struct {
-	Config configPath `type:"existingfile" help:"path to config file"`
-}
-
-func (self mineCmd) Run() error {
-	logger := logging.NewLogger()
-
-	cfg, err := config.ParseConfig(logger, string(self.Config))
-	if err != nil {
-		return errors.Wrap(err, "creating config")
-	}
-
-	// Defining a global context for starting and stopping of components.
-	ctx := context.Background()
-
-	client, err := ethereum.NewClient(ctx, logger)
-	if err != nil {
-		return errors.Wrap(err, "creating ethereum client")
-	}
-
-	accounts, err := ethereum.GetAccounts()
-	if err != nil {
-		return errors.Wrap(err, "getting accounts")
-	}
-
-	// We define our run groups here.
-	var g run.Group
-	// Run groups.
-	{
-		// Handle interupts.
-		g.Add(run.SignalHandler(context.Background(), syscall.SIGHUP, syscall.SIGINT, syscall.SIGTERM))
-
-		// Open a local or remote instance of the TSDB database.
-		var tsDB storage.SampleAndChunkQueryable
-		if cfg.Db.RemoteHost != "" {
-			tsDB, err = db.NewRemoteDB(cfg.Db)
-			if err != nil {
-				return errors.Wrap(err, "opening remote tsdb DB")
-			}
-			level.Info(logger).Log("msg", "connected to remote db", "host", cfg.Db.RemoteHost, "port", cfg.Db.RemotePort)
-		} else {
-			// Open the TSDB database.
-			tsdbOptions := tsdb.DefaultOptions()
-			// 5 days are enough as the aggregator needs data only 24 hours in the past.
-			tsdbOptions.RetentionDuration = int64(5 * 24 * time.Hour)
-			_tsDB, err := tsdb.Open(cfg.Db.Path, nil, nil, tsdbOptions)
-			if err != nil {
-				return errors.Wrap(err, "opening local tsdb DB")
-			}
-			defer func() {
-				if err := _tsDB.Close(); err != nil {
-					level.Error(logger).Log("msg", "closing the tsdb", "err", err)
-				}
-			}()
-			tsDB = _tsDB
-			level.Info(logger).Log("msg", "opened local db", "path", cfg.Db.Path)
-			level.Warn(logger).Log("msg", "FOR NEW DB INSTANCES IT IS NORMAL TO SEE SOME QUERY ERRORS AS THE DATABASE IS NOT YET POPULATED WITH VALUES")
-		}
-
-		// Web/Api server.
-		{
-			srv, err := web.New(logger, ctx, tsDB, cfg.Web)
-			if err != nil {
-				return errors.Wrap(err, "create web server")
-			}
-			g.Add(func() error {
-				err := srv.Start()
-				level.Info(logger).Log("msg", "web server shutdown complete")
-				return err
-			}, func(error) {
-				srv.Stop()
-			})
-		}
-
-		// Aggregator.
-		aggregator, err := aggregator.New(logger, ctx, cfg.Aggregator, tsDB)
-		if err != nil {
-			return errors.Wrap(err, "creating aggregator")
-		}
-
-		// Index tracker.
-		// Run only when not using remote DB as it needs to write to the local db.
-		if cfg.Db.RemoteHost == "" {
-			_tsDB, ok := tsDB.(*tsdb.DB)
-			if !ok {
-				return errors.New("tsdb is not a writable DB instance")
-			}
-
-			// Index Tracker.
-			index, err := index.New(logger, ctx, cfg.IndexTracker, _tsDB, client)
-			if err != nil {
-				return errors.Wrapf(err, "creating index tracker")
-			}
-
-			g.Add(func() error {
-				err := index.Run()
-				level.Info(logger).Log("msg", "index shutdown complete")
-				return err
-			}, func(error) {
-				index.Stop()
-			})
-
-			_netID, err := client.NetworkID(ctx)
-			if err != nil {
-				return errors.Wrap(err, "getting network ID")
-			}
-			netID := _netID.Int64()
-
-			// Dispute tracker.
-			// Run it only when not connected to a remote DB.
-			// A remote DB already runs a dispute tracker so no need to run another one.
-			// Also run and only for mainnet or rinkeby as the tellor oracle exists only on those networks.
-			if netID == 1 || netID == 4 {
-				contractTellor, err := contracts.NewITellor(client)
-				if err != nil {
-					return errors.Wrap(err, "create tellor contract instance")
-				}
-
-				disputeTracker, err := dispute.New(
-					logger,
-					ctx,
-					cfg.DisputeTracker,
-					_tsDB,
-					client,
-					contractTellor,
-					psrTellor.New(logger, cfg.PsrTellor, aggregator),
-				)
-				if err != nil {
-					return errors.Wrap(err, "creating profit tracker")
-				}
-				g.Add(func() error {
-					disputeTracker.Start()
-					level.Info(logger).Log("msg", "dispute tracker shutdown complete")
-					return nil
-				}, func(error) {
-					disputeTracker.Stop()
-				})
-			}
-
-		}
-
-		gasPriceQuerier, err := gasStation.New(logger, cfg.GasStation, client)
-		if err != nil {
-			return errors.Wrap(err, "creating gas price tracker")
-		}
-
-		if cfg.SubmitterTellor.Enabled {
-			// Profit tracker.
-			var accountAddrs []common.Address
-			for _, acc := range accounts {
-				accountAddrs = append(accountAddrs, acc.Address)
-			}
-
-			contractTellor, err := contracts.NewITellor(client)
-			if err != nil {
-				return errors.Wrap(err, "create tellor contract instance")
-			}
-
-			profitTracker, err := profit.NewProfitTracker(logger, ctx, cfg.ProfitTracker, client, contractTellor, accountAddrs)
-			if err != nil {
-				return errors.Wrap(err, "creating profit tracker")
-			}
-			g.Add(func() error {
-				err := profitTracker.Start()
-				level.Info(logger).Log("msg", "profit tracker shutdown complete")
-				return err
-			}, func(error) {
-				profitTracker.Stop()
-			})
-
-			// Event tasker.
-			tasker, taskerChs, err := tasker.New(ctx, logger, cfg.Tasker, client, contractTellor, accounts)
-			if err != nil {
-				return errors.Wrap(err, "creating tasker")
-			}
-			g.Add(func() error {
-				err := tasker.Start()
-				level.Info(logger).Log("msg", "tasker shutdown complete")
-				return err
-			}, func(error) {
-				tasker.Stop()
-			})
-
-			// Create a submitter for each account.
-			for _, account := range accounts {
-				loggerWithAddr := log.With(logger, "addr", account.Address.String()[:6])
-
-				transactor, err := transactor.New(loggerWithAddr, cfg.Transactor, gasPriceQuerier, client, account)
-				if err != nil {
-					return errors.Wrap(err, "creating transactor")
-				}
-
-				psr := psrTellor.New(loggerWithAddr, cfg.PsrTellor, aggregator)
-
-				// Get a channel on which it listens for new data to submit.
-				submitter, submitterCh, err := tellor.New(
-					ctx,
-					loggerWithAddr,
-					cfg.SubmitterTellor,
-					client,
-					contractTellor,
-					account,
-					reward.New(loggerWithAddr, aggregator, contractTellor),
-					transactor,
-					gasPriceQuerier,
-					psr,
-				)
-				if err != nil {
-					return errors.Wrap(err, "creating tellor submitter")
-				}
-				g.Add(func() error {
-					err := submitter.Start()
-					level.Info(loggerWithAddr).Log("msg", "tellor submitter shutdown complete")
-					return err
-				}, func(error) {
-					submitter.Stop()
-				})
-
-				// Will be used to cancel pending submissions.
-				tasker.AddSubmitCanceler(submitter)
-
-				// The Miner component.
-				miner, err := mining.NewMiningManager(loggerWithAddr, ctx, cfg.Mining, contractTellor, taskerChs[account.Address.String()], submitterCh, client)
-				if err != nil {
-					return errors.Wrap(err, "creating miner")
-				}
-				g.Add(func() error {
-					err := miner.Start()
-					level.Info(loggerWithAddr).Log("msg", "miner shutdown complete")
-					return err
-				}, func(error) {
-					miner.Stop()
-				})
-			}
-		}
-
-		if cfg.SubmitterTellorMesosphere.Enabled {
-			contract, err := contracts.NewITellorMesosphere(client)
-			if err != nil {
-				return errors.Wrap(err, "create contract instance")
-			}
-
-			// Create a submitter for each account.
-			for _, account := range accounts {
-				loggerWithAddr := log.With(logger, "addr", account.Address.String()[:6])
-				psr := psrTellorMesosphere.New(loggerWithAddr, cfg.PsrTellorMesosphere, aggregator)
-				transactor, err := transactor.New(loggerWithAddr, cfg.Transactor, gasPriceQuerier, client, account)
-				if err != nil {
-					return errors.Wrap(err, "creating transactor")
-				}
-
-				submitter, err := tellorMesosphere.New(
-					ctx,
-					loggerWithAddr,
-					cfg.SubmitterTellorMesosphere,
-					client,
-					contract,
-					account,
-					transactor,
-					psr,
-				)
-				if err != nil {
-					return errors.Wrap(err, "creating tellor mesosphere submitter")
-				}
-				g.Add(func() error {
-					err := submitter.Start()
-					level.Info(loggerWithAddr).Log("msg", "tellor mesosphere submitter shutdown complete")
-					return err
-				}, func(error) {
-					submitter.Stop()
-				})
-			}
-		}
-
-	}
-
-	if err := g.Run(); err != nil {
-		level.Error(logger).Log("msg", "main exited with error", "err", err)
-		return err
-	}
-
-	level.Info(logger).Log("msg", "main shutdown complete")
 	return nil
 }
