+++ pkg/format/format.go
@@ -5,8 +5,6 @@ package format
 
 import (
 	"encoding/json"
-	"math/big"
-	"strconv"
 	"strings"
 	"time"
 
@@ -47,14 +45,4 @@ func (d *Duration) UnmarshalJSON(b []byte) error {
 
 func SanitizeMetricName(input string) string {
 	return strings.ReplaceAll(input, "/", "_")
-}
-
-func ERC20Balance(balance *big.Int) string {
-	f := 0.0
-	if balance != nil {
-		divisor := big.NewInt(1e18 / 100)
-		divisor.Div(balance, divisor)
-		f = float64(divisor.Uint64()) / 100
-	}
-	return strconv.FormatFloat(f, 'f', 2, 64)
 }
